Ticket [https://trac.osgeo.org/fdo/ticket/191]: KingOracle: Add provider to top level build/release process
git-svn-id: http://svn.osgeo.org/fdo/trunk@3615 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/KingOracle/src/Provider/c_KgOraReader.h
@@ -250,7 +250,7 @@ template <class FDO_READER> FdoDateTime c_KgOraReader<FDO_READER>::GetDateTime(F
       fdodate.day = day;
       fdodate.hour = hour;
       fdodate.minute = min;
-      fdodate.seconds = sec;
+      fdodate.seconds = (float)sec;
       
     }
     return fdodate;
